backend/requirements.txt
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.16.0
packaging==21.3
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
